requirements.txt
pandas==2.1.4
PyYAML==6.0.1
numpy==1.26.4
